feat(Chest): re-work
