+++ src/main/java/github/xnzvl/karak/items/Item.java
@@ -1,6 +1,5 @@
 package github.xnzvl.karak.items;
 
-import github.xnzvl.karak.items.chests.Chest;
 import github.xnzvl.karak.items.spells.Spell;
 import github.xnzvl.karak.utils.Either;
