--- src/main/java/github/xnzvl/karak/items/chests/impl/OrdinaryChest.java
@@ -1,21 +1,0 @@
-package github.xnzvl.karak.items.chests.impl;
-
-import github.xnzvl.karak.items.chests.VariousChest;
-
-/**
- * Ordinary chest worth 1 point.
- *
- * @author Jakub Nezval
- */
-public class OrdinaryChest extends VariousChest {
-    public OrdinaryChest(
-            boolean isLocked
-    ) {
-        super(
-                "Chest",
-                "Ordinary chest worth 1 point.",
-                1,
-                isLocked
-        );
-    }
-}
