--- src/main/java/github/xnzvl/karak/items/chests/impl/RubyChest.java
@@ -1,19 +1,0 @@
-package github.xnzvl.karak.items.chests.impl;
-
-import github.xnzvl.karak.items.chests.VariousChest;
-
-/**
- * Ruby chest worth 1.5 points.
- *
- * @author Jakub Nezval
- */
-public class RubyChest extends VariousChest {
-    public RubyChest() {
-        super(
-                "Ruby Chest",
-                "Ruby chest worth 1.5 points.",
-                1.5F,
-                false
-        );
-    }
-}
